mise à jour des vues par role

+++ resources/views/classes.blade.php
@@ -9,7 +9,7 @@
 		<div class="col-md-12">
 			<nav aria-label="breadcrumb" class="float-right mt-1">
 				<ol class="breadcrumb">
-					<li class="breadcrumb-item"><a href="/accueil">Accueil</a></li>
+					<li class="breadcrumb-item"><a href="/">Accueil</a></li>
 					<li class="breadcrumb-item active" aria-current="page"><a>Classes</a></li>
 				</ol>
 			</nav>
@@ -21,7 +21,7 @@
 				<div class="card-body">
 					<div class="titre-ajout">
 						<h4 class="header-title mt-0 mb-1">Liste des classes</h4>
-						@if (auth()->user()->role != "Visiteur")
+						@if (auth()->check() && auth()->user()->role == "Administrateur")
 							<a class="ml-auto btn btn-primary" href="/classes/create">Ajouter</a>
 						@endif
 					</div>
@@ -35,7 +35,9 @@
 									<th scope="col">Filière</th>
 									<th scope="col">Ecole</th>
 									<th scope="col">Année scolaire</th>
-									<th scope="col">Actions</th>
+									@if (auth()->check() && auth()->user()->role == "Administrateur")
+										<th scope="col">Actions</th>
+									@endif
 								</tr>
 							</thead>
 							<tbody>
@@ -47,16 +49,16 @@
 										<td>{{$classe->filiere}}</td>
 										<td>{{$classe->ecole}}</td>
 										<td>{{$classe->annee_scolaire}}</td>
-										<td>
-											@if (auth()->user()->role == "Administrateur")
+										@if (auth()->check() && auth()->user()->role == "Administrateur")
+											<td>
 												<a href="/classes/{{$classe->id}}/edit" class="btn btn-primary btn-sm">
 													<i class="bi bi-pencil-square"></i>
 												</a>
 												<a href="/classes/{{$classe->id}}/delete" class="btn btn-danger btn-sm suppression">
 													<i class="bi bi-trash"></i>
 												</a>
-											@endif
-										</td>
+											</td>
+										@endif
 									</tr>
 								@endforeach
 							</tbody>
