+++ resources/views/cours.blade.php
@@ -9,7 +9,7 @@
 		<div class="col-md-12">
 			<nav aria-label="breadcrumb" class="float-right mt-1">
 				<ol class="breadcrumb">
-					<li class="breadcrumb-item"><a href="/accueil">Accueil</a></li>
+					<li class="breadcrumb-item"><a href="/">Accueil</a></li>
 					<li class="breadcrumb-item active" aria-current="page"><a>Cours</a></li>
 				</ol>
 			</nav>
@@ -21,7 +21,7 @@
 				<div class="card-body">
 					<div class="titre-ajout">
 						<h4 class="header-title mt-0 mb-1">Liste des cours</h4>
-						@if (auth()->user()->role != "Visiteur")
+						@if (auth()->check() && auth()->user()->role == "Administrateur")
 							<a class="ml-auto btn btn-primary" href="/cours/create">Ajouter</a>
 						@endif
 					</div>
@@ -37,7 +37,9 @@
 									<th scope="col">Date</th>
 									<th scope="col">Début</th>
 									<th scope="col">Fin</th>
-									<th scope="col">Action</th>
+									@if (auth()->check() && auth()->user()->role == "Administrateur")
+										<th scope="col">Action</th>
+									@endif
 								</tr>
 							</thead>
 							<tbody>
@@ -60,16 +62,16 @@
 										<td>{{$cour->date_cours}}</td>
 										<td>{{$cour->heure_debut}}</td>
 										<td>{{$cour->heure_fin}}</td>
-										<td>
-											@if (auth()->user()->role == "Administrateur")
+										@if (auth()->check() && auth()->user()->role == "Administrateur")
+											<td>
 												<a href="/cours/{{$cour->id}}/edit" class="btn btn-primary btn-sm">
 													<i class="bi bi-pencil-square"></i>
 												</a>
 												<a href="/cours/{{$cour->id}}/delete" class="btn btn-danger btn-sm suppression">
 													<i class="bi bi-trash"></i>
 												</a>
-											@endif
-										</td>
+											</td>
+										@endif
 									</tr>
 								@endforeach
 							</tbody>
